feat: add minimal security, CORS, and docker dev setup
Protect admin endpoints with X-API-Key, add configurable CORS, and align app+migrations DB config. Add docker-compose + dev bootstrap scripts for repeatable local runs.

Made-with: Cursor

+++ alembic/versions/003_add_soft_delete_and_audit.py
@@ -38,7 +38,7 @@ def upgrade() -> None:
         sa.Column('table_name', mysql.VARCHAR(80), nullable=False),
         sa.Column('record_id', mysql.BIGINT(), nullable=False),
         sa.Column('action', mysql.VARCHAR(20), nullable=False),
-        sa.Column('changed_by', mysql.BIGINT(), nullable=True),
+        sa.Column('changed_by', mysql.INTEGER(), nullable=True),
         sa.Column('changed_at', mysql.DATETIME(fsp=6), nullable=False),
         sa.Column('old_values', mysql.JSON(), nullable=True),
         sa.Column('new_values', mysql.JSON(), nullable=True),
